Format: Implement most of Format
This adds most of the functionality equivalent to the Go version's
format package.

It doesn't yet implement any of the marshalling of stanzas.

Swift's namespacing is a little unusual: without making a totally
separate module for Format, the namespace is shared with everything
else, so I ended up wrapping it in a class, but making all members
static, which feels like a pretty dirty hack. I'd love to know if
there's a better way.

I also had to implement some readLine() methods for InputStream and
Data.