Format: Rework to use ByteBuffer
Swift doesn't have anything built-in to handle byte buffers, but
SwiftNIO has exactly this.

This imports SwiftNIO as a dependency and adds an extension to
ByteBuffer for a couple of useful methods to read the buffer until a
delimiter character.

A StanzaReader is added to hold a buffer and read it to extra a Stanza.

Format.parse now also uses a buffer to make it easier to read chunks by
delimiter.

One problem with StanzaReader is it takes a copy of the buffer, so in
parse we need to copy the buffer back to get updated read index values
from it.

Another issue that needs fixing later is the ByteBuffer is consuming the
entire InputStream into the buffer, rather than only advancing the
stream as it needs to read more. This fine while I'm working on this and
using small values, but this could get memory hungry with real data.