Format: Refactor
Use the third-party ExtrasBase64 package as it allows encoding to base64
without padding characters, needed by the format.

Tidy up the general organisation: use an enum instead of a class for
namespacing as we don't need to instantiate anything: using an enum is
the idiomatic way of namespacing as it prevents instantiating the type.

Add the missing encode* functions used by functions in Primitives.swift.

Use string concat for constructing the stanzas as it removes the need
for converting to/from Data constantly.

Prefer using guard over if for conditions that force throwing an error.