ByteBuffer: Fix padded zeros on resulting buffer
When consuming from an InputStream a UInt8 array is used as a buffer to
read from the stream. The array needs to be (as far as I can tell)
filled with data, rather than creating an array and calling
reserveCapacity() - which doesn't seem to populate the array when
passing to the stream's read(_:maxLength:) method.

This means the buffer is initialized with all zeros, but if reading the
stream doesn't fill the buffer it's still got all those extra zeros so
you need to be careful when reading from the buffer to only read the
number of bytes that were actually read in to it.

To work around this an intermediary Data buffer is used, writing on the
bytes that were read from the stream.