Armor: Implement Armor
This implements everything required to produce armored (PEM encoded)
output, and decode it again.

The test to re-encode is currently failing for non-obvious reasons. By
eye the data looks the same.

The writer differs a little from the Go version, particularly in the
close() method - the Go version seems to do a lot more work to format
things correctly: this one uses the native ability to add 64-char line
endings in base64-encoding, which I think is doing the right thing. I'm
not sure yet whether it's doing the right thing with line endings,
though, which again the Go version has some special handling for.

The ByteBuffer extension now returns all data in the buffer when the
delimiter can't be found, needed for the last line of a PEM encoded
input which might not have a trailing newline.