Scrypt: Replace CryptoSwift with libscrypt wrapper
CryptoSwift's implementation of Scrypt is way too slow. The unit tests
take 1-2 minutes to run and on an iPhone simulator in Xcode generating
an Scrypt identity takes more than 2 minutes.

I'm assuming this is down to being written in pure Swift and being
unoptimised.

This switches it out for a version that wraps (and embeds) the C
libscrypt library, which is able to generate an Scrypt identity in
seconds.

To not have to refactor a bunch of things I've imported the String and
Data extensions from CryptoSwift that adds a bytes property to each.